[REFACTOR] 유저로 변환하는데 필요한 정보를 담은 OAuthAgreedUser 정의, Client -> Requester로 파일명 변경
