+++ app/user/src/main/java/com/chinda/user/domain/iam/OAuthRequester.java
@@ -1,0 +1,6 @@
+package com.chinda.user.domain.iam;
+
+public interface OAuthRequester {
+    OAuthAgreedUser getOAuthAgreedUser(String authCode);
+
+}
