--- auth/src/main/java/com/auth/auth/AuthRoot.java
@@ -1,7 +1,0 @@
-package com.auth.auth;
-
-import org.springframework.context.annotation.ComponentScan;
-
-@ComponentScan(basePackageClasses = {AuthRoot.class})
-public interface AuthRoot {
-}
