--- auth/src/main/java/com/auth/auth/client/OAuthClient.java
@@ -1,10 +1,0 @@
-package com.auth.auth.client;
-
-import com.auth.auth.dto.OauthUserResponse;
-
-public interface OAuthClient {
-
-    String getAccessToken(String authCode);
-
-    OauthUserResponse getUser(String accessToken);
-}
